requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
discord.py==2.3.2
frozenlist==1.4.0
greenlet==3.0.1
idna==3.6
multidict==6.0.4
SQLAlchemy==2.0.23
typing_extensions==4.8.0
yarl==1.9.3

python-dotenv~=1.0.1